+++ blog/mpd-docker-2014-10-20.markdown
