+++ internal/blog/blog_test.go
@@ -1,0 +1,12 @@
+package blog
+
+import (
+	"testing"
+)
+
+func TestLoadPosts(t *testing.T) {
+	_, err := LoadPosts("../../blog")
+	if err != nil {
+		t.Fatal(err)
+	}
+}
